+++ src/Twilio/Rest/Conversations/V1/Service/Conversation/ParticipantResource.cs
@@ -216,7 +216,7 @@ namespace Twilio.Rest.Conversations.V1.Service.Conversation
                                                  string roleSid = null,
                                                  string messagingBindingProxyAddress = null,
                                                  string messagingBindingProjectedAddress = null,
-                                                 long? lastReadMessageIndex = null,
+                                                 int? lastReadMessageIndex = null,
                                                  string lastReadTimestamp = null,
                                                  ParticipantResource.WebhookEnabledTypeEnum xTwilioWebhookEnabled = null,
                                                  ITwilioRestClient client = null)
@@ -256,7 +256,7 @@ namespace Twilio.Rest.Conversations.V1.Service.Conversation
                                                                                          string roleSid = null,
                                                                                          string messagingBindingProxyAddress = null,
                                                                                          string messagingBindingProjectedAddress = null,
-                                                                                         long? lastReadMessageIndex = null,
+                                                                                         int? lastReadMessageIndex = null,
                                                                                          string lastReadTimestamp = null,
                                                                                          ParticipantResource.WebhookEnabledTypeEnum xTwilioWebhookEnabled = null,
                                                                                          ITwilioRestClient client = null)
@@ -636,7 +636,7 @@ namespace Twilio.Rest.Conversations.V1.Service.Conversation
         /// Index of last “read” message in the Conversation for the Participant.
         /// </summary>
         [JsonProperty("last_read_message_index")]
-        public long? LastReadMessageIndex { get; private set; }
+        public int? LastReadMessageIndex { get; private set; }
         /// <summary>
         /// Timestamp of last “read” message in the Conversation for the Participant.
         /// </summary>
